backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
arabic-reshaper==3.0.0
asgiref==3.9.1
asn1crypto==1.5.1
attrs==25.3.0
Authlib==1.6.3
autobahn==24.4.2
Automat==25.4.16
bandit==1.8.6
billiard==4.2.2
Brotli==1.1.0
cachetools==5.5.2
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cryptography==45.0.7
cssselect2==0.8.0
daphne==4.2.1
defusedxml==0.7.1
dj-rest-auth==7.0.1
Django==5.2.6
django-allauth==0.63.4
django-cors-headers==4.7.0
django-debug-toolbar==6.0.0
django-encrypted-model-fields==0.6.5
django-extensions==4.1
django-filter==25.1
django-jsonfield==1.4.1
django-oauth-toolkit==3.0.1
django-otp==1.6.1
django-otp-yubikey==1.1.0
django-redis==6.0.0
django-stripe-payments==2.0.0
django-test-plus==2.3.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dparse==0.6.4
filelock==3.19.1
fonttools==4.60.0
freetype-py==2.5.1
geographiclib==2.1
geopy==2.4.1
google-auth==2.40.3
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.5.2
jwcrypto==1.5.6
kombu==5.5.4
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==4.0.1
mdurl==0.1.2
msgpack==1.1.1
nltk==3.9.1
numpy==2.3.3
oauthlib==3.3.1
oscrypto==1.3.0
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psutil==7.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.28.0
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.9.2
pydantic_core==2.23.4
pydyf==0.11.0
Pygments==2.19.2
pyHanko==0.31.0
pyhanko-certvalidator==0.29.0
PyJWT==2.10.1
pyOpenSSL==25.3.0
pypdf==6.1.0
pyphen==0.17.2
pytest==8.4.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
redis==6.4.0
regex==2025.9.1
reportlab==4.4.4
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.1.0
rlPyCairo==0.4.0
rsa==4.9.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
safety==3.6.1
safety-schemas==0.0.14
service-identity==24.2.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
stevedore==5.5.0
stripe==12.5.0
svglib==1.6.0
tenacity==9.1.2
tinycss2==1.4.0
tinyhtml5==2.0.0
tomlkit==0.13.3
tqdm==4.67.1
Twisted==25.5.0
txaio==25.9.2
typer==0.17.4
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
webencodings==0.5.1
xhtml2pdf==0.2.17
YubiOTP==1.0.0
zope.interface==8.0.1
zopfli==0.2.3.post1
